ref: harmonize several concepts' naming
I want to do this change now, while the repository is not yet in shared
development. The opportunities for this scale of corrective changes get
scarce later.

Rationale for this is as follows.

1. Operation -> Event

Events and operations have so far been used interchangeably, with little
consistency. This does not promote good readability. The name
_operation_ was originally chosen as it became obvious that the name
_event_ is ambiguous in the Luke domain of forestry simulation. Events
there mean forestry operations but not e.g. forest growth. In a more
general setting for a simulator, an event is anything that acts upon the
state data of the simulation. Choosing to call everything an operation
was a compromise to facilitate communication between developers and
domain experts.

I have now chosen to forgo that compromise and call everything an event.
Operation is too generic and takes mental gymnastics to associate it
with both simulation events and domain actions. While a forest
maintenance operation certainly is an operation, forest growing is
certainly not an operation. Operation does not have a meaning in the
context of a generic simulator. Possible ambiguity between the terms
should be reconciled separately in human communication rather than in
code here.

2. Payload -> State

While payload is handy lingo for talking about anything of some
importance for a task at hand, simulations act upon a state. Let's call
it state and be done with it. There is a risk of ambiguity with forestry
domain experts who associate the concept of state into a _state year_,
which represents a time point when no forestry operations take place.
Nonetheless I claim that this ambiguity should be resolved elsewhere
than in core simulator code.

3. Event labels and events

Made a clear distinction between event names represented in a control
source and the actual events (functions). Event labels are strings that
may alias each others and eventually resolve into a label that
represents an actual event name. These labels come with associated,
likewise aliased sets of parameters. Actual events are proper functions
resolved and prepared based on event labels ans parameters.

4. Continuations

Stop referring to T=>T functions as continuations. While the simulator
does implement a kind of a continuation machine with rewindable state,
continuation as a term seems reserved very specifically to preparation
of lambda functions with the continuation function given as a parameter
rather than the atomic, stateful and chained event function processing
that we are doing.