feat: replace boost::any usage with std::string
boost::any_cast is not smart enough to handle conversions from string to float, int
In turn, this makes boost::any not serve any purpose as a type-agnostic storage.
We would need to use boost::lexical_cast in any case so we might as well use std::string directly as data storage.